Throw Errors and alter code logic
added error messaged to be thrown when missing fields are added to the databse. Added the module ID to getlecture

To prevent addition of lectures with missing information as their content. added module id parameter to get lecture to specifically get the lecture depending on which module the user is on: GIT or HTML